[Move] 옵션매니저 파일 위치 변경
